% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/importingData.R
\name{poll_Import}
\alias{poll_Import}
\title{Importing Poll Data for Use in pollVizR}
\usage{
poll_Import(choices = c("C1", "C2", "Undecided"), support, decimal = T, n,
  cI = 0.95)
}
\arguments{
\item{choices}{Vector representing candidate/party names. Defaults to "C1", "C2", and "Undecided."}

\item{support}{Vector of percentage support. No default.}

\item{decimal}{Is ``support" a decimal? If so, TRUE (default). If percentage, FALSE.}

\item{n}{Number of poll respondents. No default.}

\item{cI}{Confidence interval. Select from 0.95 (default), 0.8, 0.85, 0.9, and 0.99.}
}
\description{
0.1: This function allows you to import the data for use in pollVizR. Returns a dataframe with the choices, support, standard error, and selected confidence interval as well as an introductory return sentence.
FUTURE: This function returns a list with a dataframe in [1], n in [2], and cI in [3] (reduces work in other functions).
Run this before any other function in pollVizR.
}
\note{
Important to save as an object for pollViz.
}
\examples{
fake <- poll_Import(choices=c("A", "B", "C"), support=c(0.45, 0.45, 0.05), n=250)
}

