% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/visualizingData.R
\name{pollViz}
\alias{pollViz}
\title{Visualizing Poll Data}
\usage{
pollViz(fake, cI = 95, plotTitle, wes = F)
}
\arguments{
\item{fake}{Your dataframe returned from pollImport.}

\item{cI}{*Integer* representing the confidence interval for the y-axis label. Defaults to 95.}

\item{plotTitle}{String representing the title. No default.}

\item{wes}{FUTURE. You may pass on a string representing a Wes Anderson palette from the package ``wesanderson'' here. FUTURE}
}
\description{
This function allows you to visualize the data in pollVizR. Returns a ggplot with margins of error embedded.
}
\examples{
pollViz(fake, cI = 85, plotTitle = "Fake data are cool")
}

