% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/visualizingBarWithMoE.R
\name{pollVizBar}
\alias{pollVizBar}
\title{Visualizing Poll Data--Bar}
\usage{
pollVizBar(fake, plotTitle, cI = 95, wes = F)
}
\arguments{
\item{fake}{Your dataframe returned from poll_Import. FUTURE: List returned from poll_Import.}

\item{plotTitle}{String representing the title. No default.}

\item{cI}{*Integer* representing the confidence interval for the y-axis label. Defaults to 95.}
}
\description{
This function allows you to visualize the data in pollVizR. Returns a vertical barplot with margins of error embedded.
}
\examples{
pollViz(fake, plotTitle = "Fake data are cool")
}

